+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Edited the duration of update check from `60 days` to `24 hours`
 * Add `wait_for_http_response()` function for a more generic response check
 * Add `wait_for_docker_container()` function to wait for a docker container to be ready
+* Add a `debug` command
 * [BC Break] Remove `callable $responseChecker` parameter from `wait_for_http_status()`
 * Revise the usage of the terms `command` and `task` for consistency through code and docs.